edit champions in admin
