--- src/components/Admin/Champions/Create.js
@@ -1,11 +1,0 @@
-import React,{Component} from 'react'
-
-export default class Create extends Component{
-    render(){
-        return(
-            <div>
-                Create Champion
-            </div>
-        )
-    }
-}
